requirements.txt
requests==2.32.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.1
httpx==0.27.0
tabulate==0.9.0
yfinance==0.2.61
newsapi-python==0.2.7
ollama==0.1.5
tqdm==4.66.2
pandas-ta==0.3.14b0
scipy==1.12.0
alpha-vantage==2.3.1
finviz==1.4.1
